Started Implementing the GUI.
Renamed old main.py into CLI_runner.py which may be added as an iindividual feature later for more advanced users.

Created a project root directory main.py which will call the gui.py and run the entire application. Changed the launch to show this change.

implemented a blank window as a test. More work to be done on the GUI.